+++ src/main/java/sideproject/gugumo/api/controller/CmntController.java
@@ -8,7 +8,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import sideproject.gugumo.domain.dto.CmntDto;
-import sideproject.gugumo.domain.dto.CustomUserDetails;
+import sideproject.gugumo.domain.dto.memberDto.CustomUserDetails;
 import sideproject.gugumo.request.CreateCmntReq;
 import sideproject.gugumo.request.UpdateCmntReq;
 import sideproject.gugumo.response.ApiResponse;
