+++ src/main/java/sideproject/gugumo/api/controller/MemberApiController.java
@@ -5,8 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import sideproject.gugumo.domain.dto.*;
-import sideproject.gugumo.exception.exception.DuplicateNicknameException;
+import sideproject.gugumo.domain.dto.memberDto.*;
 import sideproject.gugumo.exception.exception.NoAuthorizationException;
 import sideproject.gugumo.exception.exception.UserNotFoundException;
 import sideproject.gugumo.response.ApiResponse;
