--- src/main/java/sideproject/gugumo/domain/dto/MemberMapper.java
@@ -1,17 +1,0 @@
-package sideproject.gugumo.domain.dto;
-
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import org.mapstruct.ReportingPolicy;
-import org.mapstruct.ValueMapping;
-import org.mapstruct.factory.Mappers;
-import sideproject.gugumo.domain.entity.Member;
-import sideproject.gugumo.domain.entity.MemberStatus;
-
-@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface MemberMapper {
-
-    MemberMapper INSTANCE = Mappers.getMapper(MemberMapper.class);
-    Member toEntity(SignUpMemberDto signUpMemberDto, MemberStatus status);
-    SignUpMemberDto toDto(Member member);
-}
