+++ src/main/java/sideproject/gugumo/domain/dto/memberDto/SignUpMemberWithEmailDto.java
@@ -1,22 +1,22 @@
-package sideproject.gugumo.domain.dto;
+package sideproject.gugumo.domain.dto.memberDto;
 
-import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
 @NoArgsConstructor
-public class SignUpMemberDto {
+public class SignUpMemberWithEmailDto {
     private String nickname;
     private String username;
     private String password;
+    private String emailAuth;
 
     @Builder
-    public SignUpMemberDto(String nickname, String username, String password) {
+    public SignUpMemberWithEmailDto(String nickname, String username, String password, String emailAuth) {
         this.nickname = nickname;
         this.username = username;
         this.password = password;
+        this.emailAuth = emailAuth;
     }
 }
