+++ src/main/java/sideproject/gugumo/filter/JwtFilter.java
@@ -9,9 +9,8 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
-import sideproject.gugumo.domain.dto.CustomUserDetails;
+import sideproject.gugumo.domain.dto.memberDto.CustomUserDetails;
 import sideproject.gugumo.domain.entity.Member;
-import sideproject.gugumo.domain.entity.MemberRole;
 import sideproject.gugumo.jwt.JwtUtil;
 
 import java.io.IOException;
