+++ src/main/java/sideproject/gugumo/service/MemberService.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sideproject.gugumo.domain.dto.MemberDto;
-import sideproject.gugumo.domain.dto.SignUpMemberDto;
-import sideproject.gugumo.domain.dto.UpdateMemberDto;
+import sideproject.gugumo.domain.dto.memberDto.MemberDto;
+import sideproject.gugumo.domain.dto.memberDto.SignUpMemberDto;
+import sideproject.gugumo.domain.dto.memberDto.UpdateMemberDto;
 import sideproject.gugumo.domain.entity.Member;
 import sideproject.gugumo.exception.exception.DuplicateEmailException;
 import sideproject.gugumo.exception.exception.DuplicateNicknameException;
